Refactor some code and add login authentication
