+++ package.json
@@ -19,7 +19,6 @@
     "bootstrap": "^5.3.5",
     "gh-pages": "^6.3.0",
     "react": "^19.0.0",
-    "react-datepicker": "^8.3.0",
     "react-dom": "^19.0.0",
     "react-redux": "^9.2.0",
     "react-router-dom": "^7.5.3"
