+++ src/assets/all.scss
@@ -20,6 +20,4 @@
 @import './styles/auth.scss';
 @import './styles/member-sidebar.scss';
 @import './styles/member-style.scss';
-@import './styles/navbar.scss';
-@import './styles/footer.scss';
 @import './styles/modal-custom.scss';
