+++ src/assets/styles/auth.scss
@@ -79,8 +79,4 @@
     font-size: 1.4rem;
     color: $coffee-grey-600;
   }
-
-  .empty-div-style {
-    min-height: 2rem;
-  }
 }
