+++ src/assets/styles/member-sidebar.scss
@@ -1,4 +1,5 @@
 $sidebar-width: 196px;
+$header-height: 74.75px;
 
 .sidebar-custom {
   min-width: $sidebar-width;
@@ -80,7 +81,80 @@ $sidebar-width: 196px;
   }
 }
 
+.member-navbar {
+  z-index: 1000;
+}
+
+.member-mobile-menu-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  z-index: 998;
+  width: 100vw;
+  height: 100vh;
+  background-color: rgba(0, 0, 0, 0.6);
+}
+
+.member-mobile-menu {
+  position: fixed;
+  top: $header-height;
+  right: 0;
+  width: 50%;
+  z-index: 999;
+  font-size: 1.6rem;
+  background: white;
+  height: calc(100% - #{$header-height});
+  padding: 8px;
+  gap: 8px;
+  box-shadow: -2px 0 10px rgba(0, 0, 0, 0.3);
+  overflow-y: auto;
+
+  ul {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+
+    li {
+      margin-bottom: 1rem;
+      a,
+      button {
+        text-decoration: none;
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        color: $coffee-primary-600;
+        background-color: $coffee-grey-000;
+        height: auto;
+        max-height: none;
+        width: 100%;
+        padding: 8px;
+        text-align: left;
+        margin-top: 8px;
+        outline: none;
+        border: none;
+
+        &:focus {
+          outline: none;
+          box-shadow: none;
+        }
+
+        &:hover {
+          background-color: $coffee-primary-300;
+          color: $coffee-primary-700;
+        }
+
+        &:active {
+          background-color: $coffee-primary-800;
+          color: $coffee-primary-000;
+        }
+      }
+    }
+  }
+}
+
+
 .sidebar-style {
+
   .member-main {
     width: 100%;
     display: flex;
@@ -91,6 +165,7 @@ $sidebar-width: 196px;
     width: 100%;
     max-width: 1000px;
   }
+
 }
 
 // 控制 body 出現 sidebar-toggle 時
