+++ src/assets/styles/modal-custom.scss
@@ -4,6 +4,7 @@ body.modal-open {
     height: 100%;
   }
   
+  // Modal 遮罩
   .custom-modal {
     background-color: rgba(0, 0, 0, 0.6);
     position: fixed;
