--- src/assets/styles/navbar.scss
@@ -1,252 +1,0 @@
-@import './font-classes';
-@import '../styles/variables';
-
-$header-height: 74.75px;
-
-// 容器寬度
-.container-width {
-  padding: 0 16px;
-
-  @media (min-width: 768px) {
-    max-width: 1296px;
-    margin: 0 auto;
-    padding: 0;
-  }
-}
-
-// 手機版選單
-.mobile-menu {
-  position: fixed;
-  top: $header-height;
-  right: 0;
-  width: 50%;
-  font-size: 1.6rem;
-  height: calc(100% - #{$header-height});
-  background-color: $coffee-grey-000;
-  z-index: 1000;
-  padding: 8px;
-  gap: 8px;
-  box-shadow: -2px 0 10px rgba(0, 0, 0, 0.3);
-  overflow-y: auto;
-
-  ul {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-
-    li {
-      margin-bottom: 1rem;
-      a,
-      button {
-        text-decoration: none;
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        color: $coffee-primary-600;
-        background-color: $coffee-grey-000;
-        height: auto;
-        max-height: none;
-        width: 100%;
-        padding: 8px;
-        text-align: left;
-        margin-top: 8px;
-        outline: none;
-        border: none;
-
-        &:focus {
-          outline: none;
-          box-shadow: none;
-        }
-
-        &:hover {
-          background-color: $coffee-primary-300;
-          color: $coffee-primary-700;
-        }
-
-        &:active {
-          background-color: $coffee-primary-800;
-          color: $coffee-primary-000;
-        }
-      }
-
-      .submenu {
-        display: flex;
-        flex-direction: column;
-        align-items: flex-end;
-        a {
-          display: inline-block;
-          padding: 8px 16px;
-          width: auto;
-        }
-      }
-    }
-  }
-}
-
-.profile-btn {
-  padding: 0 16px !important;
-
-  .mobile-profile {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.mobile-menu-overlay {
-  position: fixed;
-  top: $header-height;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  z-index: 990;
-}
-
-.hamburger-btn {
-  background: none;
-  border: none;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  padding: 8px 16px;
-  gap: 4px;
-  width: auto;
-  height: 100%;
-  position: relative;
-  z-index: 1100;
-
-  span {
-    display: block;
-    width: 18px;
-    height: 2px;
-    background-color: $coffee-primary-000;
-    border-radius: 1px;
-    transition: all 0.3s ease;
-  }
-
-  &:hover span {
-    background-color: $coffee-primary-300;
-  }
-}
-
-// 桌面版選單
-.nav-brand {
-  padding: 16px 8px;
-}
-
-.nav-logo {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-
-  .logo-icon {
-    width: 27.73px;
-    height: auto;
-  }
-  .logo-text {
-    font-size: 2.4rem;
-    line-height: 150%;
-    @extend .semibold-text;
-  }
-}
-
-.navbar-menu {
-  white-space: nowrap;
-  gap: 8px;
-
-  .nav-link-custom {
-    padding: 8px 16px !important;
-    height: auto;
-    display: flex;
-    align-items: center;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    color: $coffee-primary-000;
-
-    &:hover {
-      background-color: $coffee-primary-300;
-      color: $coffee-primary-700;
-    }
-
-    &:active {
-      background-color: $coffee-primary-800;
-      color: $coffee-primary-000;
-    }
-  }
-
-  .nav-user-toggle {
-    display: inline-flex !important;
-    align-items: center;
-    gap: 8px !important;
-    padding: 8px 16px !important;
-    height: 40px;
-
-    &:hover {
-      background-color: $coffee-primary-300;
-      color: $coffee-primary-700;
-    }
-
-    &:active {
-      background-color: $coffee-primary-800;
-      color: $coffee-primary-000;
-    }
-
-    .nav-profile {
-      width: 32px;
-      height: 32px;
-    }
-  }
-}
-
-// hover 時展開 menu
-.dropdown-toggle::after {
-  display: none;
-}
-
-.hover-dropdown {
-  position: relative;
-
-  .dropdown-menu-custom {
-    display: none;
-    position: absolute;
-    top: 100%;
-    left: 50%;
-    transform: translateX(-50%);
-    z-index: 1000;
-    background-color: white;
-    height: auto;
-    max-height: none;
-    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.15);
-    padding: 8px;
-    min-width: 112px;
-    margin-top: 8px;
-
-    &.show {
-      display: block;
-    }
-
-    &.dropdown-menu-custom-about {
-      min-width: 192px;
-    }
-
-    .dropdown-item-custom {
-      color: $coffee-primary-600;
-      font-size: 1.6rem;
-      text-decoration: none;
-      height: 48px;
-      line-height: 48px;
-      text-align: center;
-      padding: 0;
-
-      &:hover {
-        background-color: $coffee-primary-300;
-        color: $coffee-primary-700;
-      }
-
-      &:active {
-        background-color: $coffee-primary-800;
-        color: $coffee-primary-000;
-      }
-    }
-  }
-}
